made changes to get the team to allow for us to keep track of points within a trick.
The winning card is broken and therefore the counting of points within a trick is broken, I think this is due to a reset error since it works for the first trick and then starts having the issue.

I think we need to decide if we are going to use the game vars that we started with.  If we do use them I think we should go 100% them and just have everything there, if we don't use them I think we should blow that whole shit away and then reconstruct them in classes/objects.
I think we should use the game manager he set up.  I think that would be much easier.